+++ ghcide/src/Development/IDE/Plugin/Completions/Types.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GADTs              #-}
 {-# LANGUAGE OverloadedLabels   #-}
+{-# LANGUAGE TypeFamilies       #-}
 module Development.IDE.Plugin.Completions.Types (
   module Development.IDE.Plugin.Completions.Types
 ) where
@@ -11,8 +12,11 @@ import qualified Data.Map                     as Map
 import qualified Data.Text                    as T
 
 import           Data.Aeson                   (FromJSON, ToJSON)
+import           Data.Hashable                (Hashable)
 import           Data.Text                    (Text)
+import           Data.Typeable                (Typeable)
 import           Development.IDE.GHC.Compat
+import           Development.IDE.Graph        (RuleResult)
 import           Development.IDE.Spans.Common
 import           GHC.Generics                 (Generic)
 import           Ide.Plugin.Config            (Config)
@@ -22,6 +26,20 @@ import           Ide.PluginUtils              (getClientConfig, usePropertyLsp)
 import           Ide.Types                    (PluginId)
 import           Language.LSP.Server          (MonadLsp)
 import           Language.LSP.Types           (CompletionItemKind (..), Uri)
+
+-- | Produce completions info for a file
+type instance RuleResult LocalCompletions = CachedCompletions
+type instance RuleResult NonLocalCompletions = CachedCompletions
+
+data LocalCompletions = LocalCompletions
+    deriving (Eq, Show, Typeable, Generic)
+instance Hashable LocalCompletions
+instance NFData   LocalCompletions
+
+data NonLocalCompletions = NonLocalCompletions
+    deriving (Eq, Show, Typeable, Generic)
+instance Hashable NonLocalCompletions
+instance NFData   NonLocalCompletions
 
 -- From haskell-ide-engine/src/Haskell/Ide/Engine/LSP/Completions.hs
